Use 60Hz for monitor on desktop
Though the monitor can support 144Hz, and does normally, when waking
from sleep it complains. It can be properly reset by using swaymsg to
change it to 60Hz and then back to 144Hz, but that's too much work. I
only need it to be 144Hz in windows for gaming anyway.